pyproject.toml
[project]
name = "simply"
version = "0.3.7"
description = "Simply: Minimal Code for Frontier LLM Research in JAX"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.3.1",
    "datasets",
    "deprecated",
    "einops>=0.8.1",
    "grain>=0.2.15",
    "orbax>=0.1.9",
    "sentencepiece",
    "setuptools<81",
    "tensorboard>=2.20.0",
    "tensorboardX",
    "jax",
    "tokenizers",
    "wandb>=0.25.0",
    "tensorflow-datasets>=4.9.9",
]


[project.optional-dependencies]
gpu = [
    "jax[cuda13]; sys_platform == 'linux'",
]
tpu = [
    "jax[tpu]>=0.8.2; sys_platform == 'linux'",
]
tfds = [
    "tensorflow-cpu; sys_platform != 'darwin'",
    "tensorflow; sys_platform == 'darwin'",
    "tensorflow_datasets",
]
gcloud = [
    "gcsfs",
]
math-eval = [
    "sympy",
    "pylatexenc",
]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["simply", "simply.*"]
